video_server/requirements.txt
aiodns
aiohttp
aioredis
cchardet
Pillow